separate pagination in its own component
